FINAL/requirements.txt
Flask==0.10.1
Flask-Admin==1.0.8
Flask-Login==0.2.11
Flask-Mail==0.9.0
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Flask-Testing==0.4.1
Flask-WTF==0.10.2
Jinja2==2.7.3
Mako==1.0.0
MarkupSafe==0.23
MySQL-python==1.2.5
Pillow==2.5.1
SQLAlchemy==0.9.7
WTForms==2.0.1
Werkzeug==0.9.6
alembic==0.6.7
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.18
blinker==1.3
boto==2.32.1
celery==3.1.13
facebook-sdk==0.4.0
itsdangerous==0.24
kombu==3.0.22
nose==1.3.3
py-bcrypt==0.4
pytz==2014.7
redis==2.10.3
simplejson==3.5.3
wsgiref==0.1.2
gunicorn==0.17.2